Added uuid in proforma invoices

+++ app/Livewire/EditProforma.php
@@ -59,8 +59,8 @@ class EditProforma extends Component
     public function addCharges($data) {
         $this->otherCharges = $data;
         if($this->otherCharges['is_taxable']) {
-            $other_charge_amount = (int) $this->otherCharges['other_charge_amount'];
-            $gst_percentage = (int) $this->otherCharges['gst_percentage'];    
+            $other_charge_amount = (float) $this->otherCharges['other_charge_amount'];
+            $gst_percentage = (float) $this->otherCharges['gst_percentage'];    
             $gst_amount = ($other_charge_amount * $gst_percentage) / (100);
             $this->otherCharges['gst_amount'] = $gst_amount;
         }
@@ -89,6 +89,7 @@ class EditProforma extends Component
     }
 
     public function mount(MakeProformaInvoice $invoice) {
+
         $invoice->load(['otherCharge', 'customer', 'terms','proformaInvoiceProducts' => function ($query) {
             $query->orderBy('sort_order', 'asc');
         }]);
@@ -127,10 +128,10 @@ class EditProforma extends Component
     public function calculateTotal() {
         $this->totalAmount = 0;
         foreach($this->addedProducts as $product) {
-            $this->totalAmount += (int) $product['quantity'] * (int) $product['price'];
+            $this->totalAmount += (float) $product['quantity'] * (float) $product['price'];
         }
         if($this->otherCharges) {
-            $this->totalAmount += (int) $this->otherCharges['other_charge_amount'];
+            $this->totalAmount += (float) $this->otherCharges['other_charge_amount'];
             if($this->otherCharges['is_taxable']) {
                 $this->totalAmount += $this->otherCharges['gst_amount'];
             }
